refactor: change interval tree fields from int to uintptr_t to represent memory addresses

+++ lib/interval-tree.h
@@ -1,21 +1,22 @@
 #ifndef INTERVAL_TREE_INT_H
 #define INTERVAL_TREE_INT_H
 
+#include <stdint.h>
 
 typedef struct Interval {
-  int low, high;
+  uintptr_t low, high;
 } Interval;
 
 typedef struct Node {
   Interval *i;
-  int max;
+  uintptr_t max;
   int height;
 
   struct Node *right, *left;
 } Node;
 
 Node *createNode(Interval *i);
-int getMaxValue(int x, int y);
+uintptr_t getMaxValue(uintptr_t x, uintptr_t y);
 void updateMax(Node *node);
 int getNodeHeight(Node *node);
 int getBalanceFactor(Node *node);
@@ -24,7 +25,7 @@ void leftRotation(Node **n);
 void insert(Node **root, Interval *i);
 void printTree(Node *root, int level);
 Node *findInterval(Node *node, Interval *i);
-Node *findPoint(Node *node, int point);
+Node *findPoint(Node *node, uintptr_t point);
 Node *getMinValue(Node *node);
 void removeInterval(Node **node, Interval *i);
